backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.1.0
numpy>=1.26.0
stable-baselines3>=2.0.0
pulp>=2.7.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
gymnasium>=0.29.0
torch>=2.0.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0